Hopfield/requirements.txt
numpy
matplotlib
pillow
alive-progress
